remove constraints covered by scalePropertiesSupportedByScaleType PR
